*** Build For Android.
The rightHeader button in Android is not responsive, so we made some change to make it more sensitive. But the final solution to make it sensible still TBD. This is just a trying.
navigation.setOptions in surveyMax index.tsx has problem in Android, the button is not easy to press. We have two methods implemented for ChatRoomSettings and AddChatRoom. Remember to find the final solution. Both are not perfect