support delete messages before a date, including the date.
still has problem on sent messages. If user sent messages, then directly go to settings to delete these messages. when back to chat room, these messages will be fetched again because the lastFetchedTimestamp is not updated in the onSend, the createdAt field is empty, we need to createdAt feedback from server