Upgrade from Expo 52 to Expo 53
We need to upgrade because iOS build fail after update the npm.
After upgrade, we need to rebuild all, please follow the document to build
Android haven't test on this package.json, so Android might not working, need test
At this checkin, the push notification is using the development aps. see entitlements
info.plist upgraded, this file is important
scheme change to "surveymax" instead of "surveyMax"